requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.0
pydantic==2.10.2
python-multipart==0.0.17

# Optional: swap SQLite for Postgres in production
# asyncpg==0.30.0
# psycopg[binary,pool]==3.2.3
